clean up primitive array module
