+++ vortex-array/src/arrays/primitive/array/patch.rs
