+++ vortex-array/src/arrays/primitive/array/top_value.rs
