+++ vortex-array/src/arrays/primitive/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::{PrimitiveArray, PrimitiveVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<PrimitiveVTable> for PrimitiveVTable {
+    fn len(array: &PrimitiveArray) -> usize {
+        array.byte_buffer().len() / array.ptype().byte_width()
+    }
+
+    fn dtype(array: &PrimitiveArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &PrimitiveArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
