+++ vortex-array/src/arrays/primitive/vtable/operations.rs
