+++ vortex-array/src/arrays/primitive/vtable/pipeline.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 use std::any::Any;
 use std::fmt::Formatter;
